refactor(notebook-cloud): own hosted-workstation state in a single hook
NotebookViewer scattered workstation state across three useState slots, a
refresh callback, an attach/default mutation pair, a cadence-driven polling
effect, and an attach-completion effect. Move all of it into
useCloudWorkstationManager, which takes capabilities plus a panel-open signal
and returns the rail/toolbar projections with the gated callbacks resolved.
notebook-viewer.tsx drops ~160 lines; workstation state now lives in one place.
