+++ apps/notebook-cloud/test/viewer-source-corpus.ts
@@ -24,6 +24,7 @@ const VIEWER_MODULE_FILES = [
   "cloud-notebook-title.tsx",
   "cloud-auth-controls.tsx",
   "use-cloud-auth.ts",
+  "use-cloud-workstations.ts",
   "cloud-viewer-config.ts",
   "cloud-viewer-types.ts",
 ] as const;
